requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
aiofiles>=23.0.0
PyJWT>=2.8.0
requests>=2.31.0
